requirements.txt
Flask
dnspython
gunicorn
aiohttp
asyncio